fix: Complete financial precision fixes for 4 more ML files
Fixed float/double → decimal conversions in:
- RiskMeasures.cs: Fixed remaining 1 double cast
- SARIPortfolioOptimizer.cs: Fixed 1 Math.Exp → DecimalMath.Exp
- IMLModel.cs: Fixed 2 Dictionary<string,double> interfaces
- MarketDataLoader.cs: Fixed 2 method parameters

Progress: 15/141 ML files fixed (11% complete)

Started work on CorrelationAnalyzer.cs but paused due to complexity
(183 occurrences, heavy MathNet.Numerics usage)

+++ DayTradinPlatform/TradingPlatform.ML/Algorithms/RAPM/RiskMeasures.cs
@@ -315,7 +315,7 @@ namespace TradingPlatform.ML.Algorithms.RAPM
             LogDebug($"Calculating Historical VaR with {returns.Length} returns");
             
             var sortedReturns = returns.OrderBy(r => r).ToArray();
-            int index = (int)Math.Floor((double)((1 - confidenceLevel) * returns.Length));
+            int index = (int)Math.Floor((1 - confidenceLevel) * returns.Length);
             
             LogDebug($"VaR index: {index}, sorted returns range: [{sortedReturns.First():F4}, {sortedReturns.Last():F4}]");
